--- a/changelogs/fragments/197-acl-fix-performance.yml
+++ b/changelogs/fragments/197-acl-fix-performance.yml
@@ -1,0 +1,2 @@
+bugfixes:
+  - "acl - Fix acl module performance (200~https://github.com/ansible-collections/ansible.posix/pull/197)."